README: update for new version
Looks like this action runner is not using "long-running branches", which means that to use the latest version of this action runner, the usage instructions need to be updated.

Alternatively, please consider having a `v1` tag which moves on every release within the `1.x` range or have a `v1` branch for the whole `1.x` development cycle.
This would mean that people would automatically get the latest version and don't have to update their workflows.